FIX GUI error message for exceptions outside the main thread
If an Exception occured outside the main thread, then the app crashed hard and didn't show any error message to the user.
